En-passant squares only when realizable
Only include the en-passant square when en-passant capture is possible
in the next move.

This effectively converts the FENs to the newer X-FEN.

The advantage is that identical position reached by different moves
yield the same FEN in this notation. For example
    1. e4 e5 2. Nf3
has the same X-FEN as
    1. Nf3 e5 2. e4
where the choice in white's move order is irrelevant; black can not
capture en-passant in any case.

For engines this is of further importance when trying to detect
repetitions/transpositions by means including an internal en-passant
square. Such a mechanism must use X-FEN semantics for compatibility with
FIDE repetition rules. These changes allow mapping such an
implementation to X-FEN notation more directly.